too_many_lines_in_file: use u32 for line count/threshold
Matches other span-offset fields in this file which are already u32;
no source file needs u64 to count its lines.

+++ clippy_config/src/conf.rs
@@ -914,7 +914,7 @@ define_Conf! {
     too_many_arguments_threshold: u64 = 7,
     /// The maximum number of lines a source file can have
     #[lints(too_many_lines_in_file)]
-    too_many_lines_in_file_threshold: u64 = 1000,
+    too_many_lines_in_file_threshold: u32 = 1000,
     /// The maximum number of lines a function or method can have
     #[lints(too_many_lines)]
     too_many_lines_threshold: u64 = 100,
